+++ ChangeLog.md
@@ -13,6 +13,8 @@ More detailed information about incremental changes can be found in the
 - Adapted to change of integer types in Spral interface. Minimal required Spral version is now v2023.03.29.
 - Fixed that return code from TNLP::eval_jac_g() was ignored at first call [#697, by
   Christoph Hansknecht].
+- Print additional messages when reallocation of MA27 working space failed
+  [#671, by Daniel Oliveira].
 
 ### 3.14.12 (2023-04-05)