+++ ChangeLog.md
@@ -7,6 +7,8 @@ More detailed information about incremental changes can be found in the
 
 ## 3.14
 
+### 3.14.14 (202x-yy-zz)
+
 ### 3.14.13 (2023-11-08)
 
 - Reduced priority for making Spral the default value for option linear_solver [#677].